Résulotion de quelques warnings

+++ src/sched.h
@@ -12,12 +12,12 @@ enum pcb_state_e {READY, NEW, TERMINATED, WAITING};
 struct pcb_s {
   /* Pointer to stack */
   uint32_t* sp;
-    
+
   /* function and args */
   func_t* entry_point;
   void* args;
-  
-  unsigned int size;
+
+  size_t size;
   char* stack_base;
   enum pcb_state_e state;
 
@@ -25,7 +25,7 @@ struct pcb_s {
 };
 
 
-int create_process(func_t* f, unsigned size);
+int create_process(func_t* f, size_t size);
 void yield();
 void start_sched();
 void schedule();
